+++ movie_rental_backend/src/main/java/com/rental_backend/repository/CustomerRepository.java
@@ -38,7 +38,7 @@ public interface CustomerRepository extends CrudRepository<Customer,Long> {
     @Modifying
     @Transactional
     @Query("update Customer c set c.balance=:balance where c.uId=:uId ")
-    void updateBalance(@Param("uId") Long uId, @Param("balance") Double balance);
+    void updateBalance(@Param("uId") Long uId, @Param("balance") float balance);
 
     @Query("select c.email from Customer c")
     List<String> findAllEmails();
